+++ tests/e2e/utils/CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added `deleteAllOrders()` that goes through and deletes all orders
 - Added `deleteAllShippingClasses()` which permanently deletes all shipping classes using the API
 - Added `statuses` optional parameter to `deleteAllRepositoryObjects()` to delete on specific statuses
+- Added `createOrder()` component util that creates an order using the API with the passed in details
 - Updated `addShippingZoneAndMethod` to use the API instead of UI to create shipping zones
 
 # 0.1.5